Clean up style/formatting
